agent_service/requirements.txt
annotated-doc==0.0.4
annotated-types==0.7.0
anyio==4.12.0
certifi==2025.11.12
charset-normalizer==3.4.4
click==8.3.1
colorama==0.4.6
fastapi==0.125.0
gigachat==0.1.43
grpcio==1.76.0
h11==0.16.0
h2==4.3.0
hpack==4.1.0
httpcore==1.0.9
httpx==0.28.1
hyperframe==6.1.0
idna==3.11
jsonpatch==1.33
jsonpointer==3.0.0
langchain-core==1.2.4
langchain-gigachat==0.3.12
langchain-qdrant==1.1.0
langgraph==1.0.5
langgraph-checkpoint==3.0.1
langgraph-prebuilt==1.0.5
langgraph-sdk==0.3.1
langsmith==0.5.0
numpy==2.3.5
orjson==3.11.5
ormsgpack==1.12.1
packaging==25.0
portalocker==3.2.0
protobuf==6.33.2
pydantic==2.12.5
pydantic-settings==2.12.0
pydantic_core==2.41.5
python-dotenv==1.2.1
pywin32==311
PyYAML==6.0.3
qdrant-client==1.16.2
redis==7.1.0
requests==2.32.5
requests-toolbelt==1.0.0
starlette==0.50.0
tenacity==9.1.2
types-requests==2.32.4.20250913
typing-inspection==0.4.2
typing_extensions==4.15.0
urllib3==2.6.2
uuid_utils==0.12.0
uvicorn==0.40.0
xxhash==3.6.0
zstandard==0.25.0
